Added SideBar + refactor
