+++ package.json
@@ -13,7 +13,7 @@
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@fontsource/roboto": "^5.0.13",
-    "@mui/icons-material": "^5.15.16",
+    "@mui/icons-material": "^5.15.18",
     "@mui/material": "^5.15.16",
     "@vis.gl/react-google-maps": "^0.9.0",
     "react": "^18.2.0",
